Created new used interface class.
This class is for anything dealing with what the user see at and the
work flow of the social network. However, it does not deal with the
behind-the-scenes logic for any of the actions. I believe that will be
done in methods of the client class which the user interface class will
call.